% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_andamento_trf2.R
\name{baixar_andamento_trf2}
\alias{baixar_andamento_trf2}
\title{Baixar movimentação processual}
\usage{
baixar_andamento_trf2(codigo_processo = NULL, diretorio = ".")
}
\arguments{
\item{codigo_processo}{Código do processo}

\item{diretorio}{Se não informado, atual}
}
\value{
html
}
\description{
Baixar movimentação processual
}
\examples{
\dontrun{
baixar_andamento_tf2(codigo_processo = 964285, diretorio = ".")
}
}
