% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_julgados_trf2.R
\name{baixar_julgados_trf2}
\alias{baixar_julgados_trf2}
\title{Baixar inteiro teor do acórdão}
\usage{
baixar_julgados_trf2(urls, diretorio = ".")
}
\arguments{
\item{urls}{As urls podem ser obtidas da tibble lida
por meio da função ler_cjsg_trf2}

\item{diretorio}{atual se não informado}
}
\value{
baixa pdfs no diretorio indicado
}
\description{
Baixar inteiro teor do acórdão
}
\examples{
\dontrun{
url <- "http://jurisprudencia.trf2.jus.br/sm/download?name=apolo-inteiro-teor&id=2019,08,02,00084163220184020000_2257151.pdf"
baixar_julgados_trf2(url)
}
}
