% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjsg_trf2.R
\name{ler_cjsg_trf2}
\alias{ler_cjsg_trf2}
\title{Lê metadados do resultado de baixar_cjsg_trf2}
\usage{
ler_cjsg_trf2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se arquivos for NULL, indicar diretório.}
}
\value{
Tibble com informações processuais
}
\description{
Lê metadados do resultado de baixar_cjsg_trf2
}
\examples{
\dontrun{
ler_cjsg_trf2(diretorio = ".")
}
}
