% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjsg_trf2_2.R
\name{ler_cjsg_trf2_2}
\alias{ler_cjsg_trf2_2}
\title{Lê htmls baixados por baixar_cjsg_trf2}
\usage{
ler_cjsg_trf2_2(diretorio = ".")
}
\arguments{
\item{diretorio}{Default para o atual}
}
\value{
tibble com informações processuais
}
\description{
Lê htmls baixados por baixar_cjsg_trf2
}
\examples{
\dontrun{
ler_cjsg_trf2(diretorio = ".")
}
}
