% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_julgados_trf2.R
\name{ler_julgados_trf2}
\alias{ler_julgados_trf2}
\title{Lê os julgados baixados com baixar_julgados_trf2}
\usage{
ler_julgados_trf2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se arquivos for NULL, informar diretório}
}
\value{
tibble com número do processo e julgado
}
\description{
Lê os julgados baixados com baixar_julgados_trf2
}
\examples{
\dontrun{
julgados <- ler_julgados_trf2(diretorio = "agencia/julgados")
}

}
